fix(CB-117): :bug: Recompiled stockfish manually on mac to fix error
Stockfish on MacOS previously failed to run with the version installed. I went on the regular github repo for Stockfish and manually recompiled using the Makefile and replaced the existing version with that. The values for analysis returned are different between the previous ones for Linux/Windows and the one for MacOS, so we may need to update the other two if we want the values to be consistent